requirements.txt
edge-tts>=6.1.12
imageio-ffmpeg>=0.6.0
moviepy>=2.1.2
pillow>=11.1.0
requests>=2.31.0